add validate and handle try catch block

+++ src/com/vng/zing/managementuser/services/UserService.java
@@ -5,6 +5,9 @@
  */
 package com.vng.zing.managementuser.services;
 
+import com.vng.zing.dmp.common.exception.ZInvalidParamException;
+import com.vng.zing.exception.InvalidParamException;
+import com.vng.zing.exception.NotExistException;
 import com.vng.zing.logger.ZLogger;
 import com.vng.zing.managementuser.entity.ApiResponse;
 import com.vng.zing.managementuser.entity.UserDTO;
@@ -20,12 +23,13 @@ import com.vng.zing.userservice.thrift.UpdateUserParams;
 import com.vng.zing.userservice.thrift.UpdateUserResult;
 import com.vng.zing.userservice.thrift.User;
 import com.vng.zing.userservice.thrift.wrapper.UserMwClient;
-import com.vng.zing.utils.DateTimeUtils;
-import com.vng.zing.zcommon.thrift.ECode;
+import com.vng.zing.managementuser.utils.DateTimeUtils;
+import com.vng.zing.managementuser.utils.RequestUtils;
 import org.apache.log4j.Logger;
+import org.apache.thrift.TException;
 import org.json.JSONException;
 import org.json.JSONObject;
-import org.json.simple.JSONArray;
+import org.json.simple.parser.ParseException;
 
 /**
  *
@@ -34,21 +38,22 @@ import org.json.simple.JSONArray;
 public class UserService {
 
     private static final Logger _Logger = ZLogger.getLogger(UserListService.class);
-    public static final UserService Instance = new UserService();
     public static final UserMwClient client = new UserMwClient("Main");
+    public static final ValidateService validateService = new ValidateService();
 
-    private UserService() {
+    public UserService() {
     }
 
-    public JSONObject getUser(int id) throws JSONException {
+    public Object getUser(int id) throws JSONException, TException {
         ThreadProfiler profiler = Profiler.getThreadProfiler();
-        JSONArray arrayData = new JSONArray();
         ApiResponse apiResponse = new ApiResponse();
-        try {
+        JSONObject temp = new JSONObject();
+        if (!validateService.validateId(id)) {
+            throw new ZInvalidParamException("User ID is not valid");
+        } else {
             DetailUserResult result = client.getUser(new DetailUserParams(id));
             profiler.push(this.getClass(), "output");
             apiResponse.setCode(result.getCode());
-            JSONObject temp = new JSONObject();
             if (result.getData() != null) {
                 User user = result.getData();
                 UserDTO myUserDTO = new UserDTO(user.id, user.name, user.username, user.gender, DateTimeUtils.getLocalDateTime(user.birthday), DateTimeUtils.getLocalDateTime(user.createtime), DateTimeUtils.getLocalDateTime(user.updatetime));
@@ -56,59 +61,45 @@ public class UserService {
             }
             apiResponse.setData(new JSONObject().put("user", temp));
             profiler.pop(this.getClass(), "output");
-        } catch (Exception ex) {
-            _Logger.error(null, ex);
-            apiResponse.setCode(ECode.C_FAIL.getValue());
-            apiResponse.setData(new JSONObject());
+            return apiResponse;
         }
-        return apiResponse.getFinalResponse();
     }
 
-    public JSONObject updateUser(User newUser) throws JSONException {
+    public Object updateUser(Object newUser) throws JSONException, ParseException, NotExistException, InvalidParamException, TException {
         ThreadProfiler profiler = Profiler.getThreadProfiler();
         ApiResponse apiResponse = new ApiResponse();
-        try {
-            UpdateUserResult result = client.updateUser(new UpdateUserParams(newUser));
-            apiResponse.setCode(result.getCode());
-            apiResponse.setData(new JSONObject().put("message", result.getMessage()));
-            profiler.pop(this.getClass(), "output");
-        } catch (Exception ex) {
-            _Logger.error(null, ex);
-            apiResponse.setCode(ECode.C_FAIL.getValue());
-            apiResponse.setData(new JSONObject());
-        }
-        return apiResponse.getFinalResponse();
+        validateService.validateUpdateUserParams(newUser);
+        UpdateUserResult result = client.updateUser(new UpdateUserParams(RequestUtils.parseUpdateUserParams(newUser)));
+        profiler.pop(this.getClass(), "output");
+        apiResponse.setCode(result.getCode());
+        apiResponse.setData(new JSONObject().put("message", result.getMessage()));
+        return apiResponse;
     }
 
-    public JSONObject createUser(User newUser) throws JSONException {
+    public Object createUser(Object newUser) throws JSONException, ParseException, TException, NotExistException, InvalidParamException {
         ThreadProfiler profiler = Profiler.getThreadProfiler();
         ApiResponse apiResponse = new ApiResponse();
-        try {
-            CreateUserResult result = client.createUser(new CreateUserParams(newUser));
-            apiResponse.setCode(result.getCode());
-            apiResponse.setData(new JSONObject().put("message", result.getMessage()));
-            profiler.pop(this.getClass(), "output");
-        } catch (Exception ex) {
-            _Logger.error(null, ex);
-            apiResponse.setCode(ECode.C_FAIL.getValue());
-            apiResponse.setData(new JSONObject());
-        }
-        return apiResponse.getFinalResponse();
+        validateService.validateCreateUserParams(newUser);
+        CreateUserResult result = client.createUser(new CreateUserParams(RequestUtils.parseCreateUserParams(newUser)));
+        profiler.pop(this.getClass(), "output");
+        apiResponse.setCode(result.getCode());
+        apiResponse.setData(new JSONObject().put("message", result.getMessage()));
+        return apiResponse;
     }
 
-    public JSONObject deleteUser(int id) throws JSONException {
+    public Object deleteUser(int id) throws JSONException, TException {
         ThreadProfiler profiler = Profiler.getThreadProfiler();
         ApiResponse apiResponse = new ApiResponse();
-        try {
+        JSONObject temp = new JSONObject();
+        if (!validateService.validateId(id)) {
+            throw new ZInvalidParamException("User ID is not valid");
+        } else {
             DeleteUserResult result = client.deleteUser(new DeleteUserParams(id));
+            profiler.push(this.getClass(), "output");
             apiResponse.setCode(result.getCode());
             apiResponse.setData(new JSONObject().put("message", result.getMessage()));
             profiler.pop(this.getClass(), "output");
-        } catch (Exception ex) {
-            _Logger.error(null, ex);
-            apiResponse.setCode(ECode.C_FAIL.getValue());
-            apiResponse.setData(new JSONObject());
+            return apiResponse;
         }
-        return apiResponse.getFinalResponse();
     }
 }
